Refactor #185 fix build
